improvement: change requirements needed to use the library
Consumers no longer need to provide a logger and a pool for connections to the database. The
constructor now accepts some options to create the logger and pool.

BREAKING CHANGE: The constructor does not accept a logger and pool. Now takes an options object to
create the dependencies for the consumer.