fix: remove dependency on jbaudanza/rxnotifier
The repository seems to be unmaintained and there was what seemed to be a bug when using the library
with typescript. One of the Rxjs constructors was deprecated. This was changed and code was
converted to typescript and refactored. Thus, there is no need to depend on the library. Unit tests
were added. This issue was found when using it in another project.